+++ changelog/_6272.md
@@ -1,0 +1,3 @@
+### Changed
+
+- moved Pagination from /budgeting to /contrib
