+++ meinberlin/react/budgeting/BudgetingProposalList.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState, useRef } from 'react'
 import django from 'django'
 import { useLocation, useSearchParams } from 'react-router-dom'
 import { EndSessionLink } from './EndSessionLink'
-import { Pagination } from './Pagination'
+import { Pagination } from '../contrib/Pagination'
 import { Card } from '../contrib/Card'
 import { CardMeta } from '../contrib/CardMeta'
 import { CardStatus } from '../contrib/CardStatus'
